% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_quantile_pit.R
\name{bernstein_e}
\alias{bernstein_e}
\title{Test stochastic dominance with Bernstein polynomials}
\usage{
bernstein_e(
  z,
  n0 = 10,
  m = 20,
  crit = "CN",
  settings = list(eps_abs = 1e-05, eps_rel = 1e-05, max_iter = 4000L, verbose = FALSE),
  ...
)
}
\arguments{
\item{z}{numbers in [0, 1].}

\item{n0}{minimum number of observations for starting. All e-values until
\code{n0} (included) are equal to 1, and the e-value is computed
starting from the \code{n0+1}th observation.}

\item{m}{maximum degree of the Bernstein polynomials in the mixture. If
\code{NULL}, an optimal \code{m} is estimated with the AIC, BIC or CN
criterion specified.}

\item{crit}{the type of criterion to use for selecting the number of weights.
Defaults to \code{"CN"}, other options are \code{"AIC"} and \code{"BIC"}.}

\item{settings}{options for \code{\link[osqp]{osqp}}.}

\item{...}{currently not used, only for calling the function with
\code{do.call}.}
}
\value{
A list containing the e-values (\code{e}).
}
\description{
Tests the null hypothesis that \code{z} are generated from a distribution
stochastically greater than the uniform distribution.
}
\details{
For the reverse hypothesis (smaller than uniform distribution), replace
\code{z} by \code{1-z}. Apply \code{\link{evalue_correct}} to prevent e-values
of exactly zero.
}
\examples{
# Test hypothesis "greater than uniform distribution in stochastic dominance"

## Hypothesis is violated
z <- rbeta(300, 0.75, 1)
e <- bernstein_e(z = z)
max(cumprod(e$e))

## Hypothesis is correct
z <- rbeta(300, 1, 0.75)
e <- bernstein_e(z = z)
max(cumprod(e$e))
}
