% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_pit.R
\name{beta_e}
\alias{beta_e}
\title{Test uniform against beta distribution}
\usage{
beta_e(z, n0 = 10, tol = 1e-06, max_it = 20, ...)
}
\arguments{
\item{z}{vector of numbers in (0,1).}

\item{n0}{minimum number of observations for starting. All e-values until
\code{n0} (included) are equal to 1, and the density is estimated
starting from the \code{n0+1}th observation.}

\item{tol}{tolerance for likelihood maximization. The maximization algorithm
is stopped as soon as the difference between the log-likelihood of two
subsequent iterations is at most \code{tol}.}

\item{max_it}{maximum number of iterations for likelihood maximization.}

\item{...}{currently not used, only for calling the function with
\code{do.call}.}
}
\value{
A list containing the vector of e-values (\code{e}) and a matrix of the
estimated parameters at all time points (\code{pars}).
}
\description{
Tests the null hypothesis that \code{z} is uniformly distributed against
the alternative that it follows a beta distribution with parameters not
equal to 1.
}
\examples{
n <- 100
z <- simulate_pit(n = n, type = "pit", bias = 0, dispersion = 0.25)$z
e <- beta_e(z = z, n0 = 5)
str(e)

# Merge by product
prod(e$e)
max(cumprod(e$e))

# Look at parameters
tail(e$par)
}
