% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beta_e_cpp}
\alias{beta_e_cpp}
\title{Compute e-values based on beta distrbution}
\usage{
beta_e_cpp(z, tol, max_it, n0)
}
\arguments{
\item{z}{numbers in (0,1), should not be equal to 0 or 1.}

\item{tol}{tolerance for stopping (small positive number).}

\item{max_it}{maximum number of iterations in Newton method.}

\item{n0}{starting sample sizes. The first \code{n0} observations are
used only for estimation and the e-values are not computed.}
}
\value{
List containing the e-values (density evaluated at the given value of
\code{z[i]}) and a matrix of the parameters for each index \code{i}.
}
\description{
Sequentially estimate a beta distribution and evaluate the density at the
next point in the sample.
}
\details{
This function is intended for internal use only.
}
\keyword{internal}
