% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beta_mle_sequential}
\alias{beta_mle_sequential}
\title{Maximum likelihood estimation for beta distribution}
\arguments{
\item{tol}{tolerance for stopping (small positive number).}

\item{max_it}{maximum number of iterations in Newton method.}

\item{sly1}{sum of log-observations.}

\item{sly2}{sum of logs of 1 minus the observations.}

\item{sy}{sum of observations.}

\item{sy2}{of squared observations.}

\item{n}{sample size.}
}
\value{
NumericVector containing the parameters of the beta distribution.
}
\description{
Estimates a beta-distribution by maximum likelihood, with pre-formatted
inputs for efficiency.
}
\details{
Estimation is performed with a Newton method and the moment matching
estimators as starting values. Estimated parameters below 0.001 or above 100
are truncated for stability.

This function is intended for internal use only.
}
\author{
Code is ported to Rcpp from the Rfast package (beta.mle) by Alexander Henzi.
}
\keyword{internal}
