% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_rank_histogram.R
\name{betabinom}
\alias{betabinom}
\alias{betabinom_e}
\alias{betabinom_e_agg}
\title{Test discrete uniform distribution against betabinomial}
\usage{
betabinom_e(r, m, n0 = 20, tol = 1e-07, max_it = 20, ...)

betabinom_e_agg(r, m, avg = TRUE, n0 = 20, tol = 1e-07, max_it = 20, ...)
}
\arguments{
\item{r}{vector or matrix of observations.}

\item{m}{size of discrete uniform distribution.}

\item{n0}{minimum number of observations for starting. All e-values until
\code{n0} (included) are equal to 1.}

\item{tol}{tolerance for likelihood maximization. The maximization algorithm
is stopped as soon as the difference between subsequent parameter
estimates is smaller than \code{tol}.}

\item{max_it}{maximum number of iterations for likelihood maximization.}

\item{...}{currently not used, only for calling the function with
\code{do.call}.}

\item{avg}{logical specify whether e-values should be averaged (across rows)
when \code{r} is a matrix.}
}
\value{
A list containing the e-values (\code{e}) and the estimated parameters
(matrix, \code{pars}).
}
\description{
Tests the null hypothesis that \code{r} are generated from a discrete
uniform distribution, against a betabinomial distribution (estimated by
maximum likelihood.)
}
\details{
If \code{n0} is too small, the estimated parameters may diverge for small
samples.

If r is a matrix, estimation is performed separately for each row - the data
is aggregated across the columns. In this case, \code{n0} is the number of
rows of observations that are used for starting. This is done using the
\code{betabinom_e_agg()} function.
}
\examples{
n <- 360

# vector
r <- simulate_pit(n, "rank_histogram", K = 20, bias = 0.2, dispersion = 0)
e <- betabinom_e(r = r$r, m = 20)
max(cumprod(e$e))

# matrix
d <- 3
r <- simulate_pit(n*d, "rank_histogram", K = 20, bias = 0.2, dispersion = 0)$r
r <- matrix(r, ncol = d)
e <- betabinom_e_agg(r = r, m = 20)

}
