% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_rank_histogram.R
\name{betabinom_e}
\alias{betabinom_e}
\title{Test discrete uniform distribution against betabinomial}
\usage{
betabinom_e(r, m, n0 = 10, tol = 1e-07, max_it = 20, ...)
}
\arguments{
\item{r}{observations.}

\item{m}{size of discrete uniform distribution.}

\item{n0}{minimum number of observations for starting. All e-values until
\code{n0} (included) are equal to 1.}

\item{tol}{tolerance for likelihood maximization. The maximization algorithm
is stopped as soon as the difference between subsequent parameter
estimates is smaller than \code{tol}.}

\item{max_it}{maximum number of iterations for likelihood maximization.}

\item{...}{currently not used, only for calling the function with
\code{do.call}.}
}
\value{
A list containing the e-values (\code{e}) and the estimated parameters
(matrix, \code{pars}).
}
\description{
Tests the null hypothesis that \code{r} are generated from a discrete
uniform distribution, against a betabinomial distribution (estimated by
maximum likelihood.)
}
\details{
If \code{n0} is too small, the estimated parameters may diverge for small
samples.
}
\examples{
r <- simulate_pit(360, "rank_histogram", K = 20, bias = 0.2, dispersion = 0)
e <- betabinom_e(r = r$r, m = 20)
max(cumprod(e$e))
}
