% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{betabinom_e_cpp}
\alias{betabinom_e_cpp}
\title{Compute e-values based on betabinomial distrbution}
\usage{
betabinom_e_cpp(r, N, tol, max_it, n0)
}
\arguments{
\item{r}{numbers in 1,...,m.}

\item{N}{number of trials (equal to \code{m} above).}

\item{tol}{tolerance for stopping (small positive number).}

\item{max_it}{maximum number of iterations in Newton method.}

\item{n0}{starting sample sizes. The first \code{n0} observations are
used only for estimation and the e-values are not computed.}
}
\value{
List containing the e-values (density evaluated at the given value of
\code{r[i]}) and a matrix of the parameters for each index \code{i}.
}
\description{
Sequentially estimate a betabinomial distribution and compute e-values.
}
\details{
This function is intended for internal use only.
}
\keyword{internal}
