% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{betabinom_mle_sequential}
\alias{betabinom_mle_sequential}
\title{Maximum likelihood estimation for betabinomial distribution}
\arguments{
\item{N}{number of trials (int).}

\item{n}{total sample size (int).}

\item{z}{vector from 0 to N (NumericVector).}

\item{n1}{vector of counts how often \code{z[i]} has been observed
(NumericVector).}

\item{n2}{vector of counts how often \code{N-z[i]} has been observed
(NumericVector).}

\item{m1}{sums of observations (double).}

\item{m2}{sums of squared observations (double).}

\item{tol}{tolerance for stopping (small positive number).}

\item{max_it}{maximum number of iterations in Newton method.}
}
\value{
NumericVector containing the parameters of the betabinomial distribution.
}
\description{
Estimates a betabinomial distribution by maximum likelihood, with
pre-formatted inputs for efficiency.
}
\details{
Estimation is performed with a Newton method and the moment matching
estimators as starting values. Estimated parameters below 0.001 or above 100
are truncated for stability.

This function is intended for internal use only.
}
\author{
Code is ported to Rcpp from the Rfast package (beta.mle) by Alexander Henzi.
}
\keyword{internal}
