% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_forecasts}
\alias{check_forecasts}
\title{Check forecasts}
\usage{
check_forecasts(y, mat)
}
\arguments{
\item{y}{vector of forecasts.}

\item{mat}{matrix of forecast quantiles or ensemble forecasts.}
}
\value{
Error message if objects are not of specified types or contain NA or sizes
do not match.
}
\description{
Check forecasts
}
\keyword{internal}
