% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_h}
\alias{check_h}
\title{Check forecast lag}
\usage{
check_h(h)
}
\arguments{
\item{h}{forecast lag.}
}
\value{
Error message if \code{h} is not a single positive integerl.
}
\description{
Check forecast lag
}
\keyword{internal}
