% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ranks}
\alias{check_ranks}
\title{Check ranks}
\usage{
check_ranks(r, m)
}
\arguments{
\item{r}{ranks for rank histogram.}

\item{m}{ensemble size.}
}
\value{
Error message if \code{r} is not a vector of positive integers smaller or
equal to \code{m+1}. Warning if the maximum of \code{r} is stritcly less than
\code{m+1}.
}
\description{
Check ranks
}
\keyword{internal}
