% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_z}
\alias{check_z}
\title{Check PIT}
\usage{
check_z(z)
}
\arguments{
\item{z}{PIT or quantile PIT.}
}
\value{
Error message if \code{z} is not a vector in [0,1] without missing values.
}
\description{
Check PIT
}
\keyword{internal}
