% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_misc.R
\name{combine_e_h}
\alias{combine_e_h}
\title{Combine lag h e-values}
\usage{
combine_e_h(es)
}
\arguments{
\item{es}{list of e-values.}
}
\value{
Combine e-value.
}
\description{
Combine e-values for forecasts with lag \code{h > 1}.
}
\details{
It is assumed that the first enrty of \code{es} contains all e-values with
indices \code{1,1+h,1+2h,...}, and more general the \code{j}th entry contains
the e-values with indices \code{j,j+h,j+2h,...}.
}
