% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein_densities.R
\name{constraintMat}
\alias{constraintMat}
\title{Generate constraint matrix}
\usage{
constraintMat(m, maxplace)
}
\arguments{
\item{m}{maximal number of degrees for Bernstein polynomial.}

\item{maxplace}{location of the mode.}
}
\value{
The constraint matrix.
}
\description{
Generates constraint matrix for estimating monotone density
}
\author{
Code by Bradley C. Turnbull, NC State University, Dr. Sujit K. Ghosh, NC
State University, August 25, 2013.
}
\keyword{internal}
