% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dbetabinom}
\alias{dbetabinom}
\title{Probability function of betabinomial distribution}
\arguments{
\item{x}{quantile (single number, not vectorized).}

\item{a}{first parameter of beta distribution (single number).}

\item{b}{first parameter of beta distribution (single number).}

\item{N}{number of trials (single number).}
}
\value{
Probability of \code{x} under the distribution specified by the parameters.
}
\description{
Computes the probability function of the betabinomial distribution.
}
\keyword{internal}
