% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_rank_histogram.R
\name{e_rank_histogram}
\alias{e_rank_histogram}
\title{E-values for rank histograms}
\usage{
e_rank_histogram(
  r,
  h,
  m,
  strategy = "betabinom",
  options = list(),
  check = FALSE
)
}
\arguments{
\item{r}{observed ranks (numbers in \code{1,...,m+1}).}

\item{h}{forecast lag.}

\item{m}{size of the ensemble (positive integer, the \code{m} above).}

\item{strategy}{strategy for evaluating calibration. Available are
\code{"empirical"} for the empirical distribution, and \code{"betabinom"}
for the beta-binomial distribution (estimated with maximum likelihood).}

\item{options}{options for the given strategy (see \code{\link{betabinom_e}}
and \code{\link{empirical_e}} for the available parameters).}

\item{check}{check for correct format of input parameters.}
}
\value{
If \code{h} equals 1: Returns a list containing the vector of e-values
(\code{e}) and the forecast lag \code{h}.

If \code{h} is greater than 1: Instead of a vector of e-values, the list
contains for each \code{j=1,2,...,h} a list with the e-values for all
observations with indices \code{j,j+h,j+2h,...}.
}
\description{
Tests null hypothesis that rank histogram follows a uniform distribution.
}
\details{
It is recommended to use \code{strategy="empirical"} only for large sample
sizes (>1000).

If r is a matrix, then the rows will be treated as different time points, and
the columns as different variables to aggregate over. This accounts for cases
where more than one forecast and observation is available at each point in time,
e.g. from multiple points in space.
}
\examples{
n <- 360
r <- simulate_pit(n, "rank_histogram", K = 20, bias = 0.2, dispersion = 0)$r

# If z is a lag 1 forecast:
e <- e_rank_histogram(r, h = 1, m = 20, strategy = "betabinom")
evalue_merge(e)
prod(e$e)
max(cumprod(e$e))

# Lag 2:
e <- e_rank_histogram(r, h = 2, m = 20, strategy = "betabinom")
str(e)
evalue_merge(e)
}
\seealso{
\code{\link{ensemble_rank}} to compute \code{r} from ensemble forecasts and
observations.
}
