% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_misc.R
\name{ensemble_rank}
\alias{ensemble_rank}
\title{Ranks for rank histogram}
\usage{
ensemble_rank(y, ensemble, check = FALSE)
}
\arguments{
\item{y}{observations (vector).}

\item{ensemble}{ensemble forecasts (matrix with \code{length(y)} rows).}

\item{check}{check for correct format of input parameters.}
}
\value{
Vector of ranks.
}
\description{
Computes the rank of an observation among an ensembles of forecasts.
}
\details{
In case of ties, the ranks are randomized.
}
