% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{epit-package}
\alias{epit-package}
\title{Tests for Probability Integral Transform And Related}
\description{
Blabla
}
\section{How does it work?}{

Blublublu
}

\section{The \pkg{epit} package}{


To use the package bliblablu
\itemize{
\item blo
\item ble
}
}

\references{
Preprint here
}
