% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_misc.R
\name{evalue_correct}
\alias{evalue_correct}
\title{Correct e-value so that it never equals zero}
\usage{
evalue_correct(e)
}
\arguments{
\item{e}{numeric vector of e-values.}
}
\value{
Vector of corrected e-values.
}
\description{
Transforms an e-value \code{e} so that it is never zero, by replacing
\code{e[i]} with \code{1/i + (1-1/i)*e[i]}.
}
