% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_quantile_pit.R
\name{grenander_e}
\alias{grenander_e}
\title{Test stochastic dominance with Grenander estimator.}
\usage{
grenander_e(z, n0 = 10, ...)
}
\arguments{
\item{z}{numbers in (0, 1] (exact zeros are not allowed).}

\item{n0}{minimum number of observations for starting. All e-values until
\code{n0} (included) are equal to 1, and the e-value is computed
starting from the \code{n0+1}th observation.}

\item{...}{currently not used, only for calling the function with
\code{do.call}.}
}
\value{
A list containing the e-values (\code{e}).
}
\description{
Tests the null hypothesis that \code{z} are generated from a distribution
stochastically greater than the uniform distribution.
}
\details{
For the reverse hypothesis (smaller than uniform distribution), replace
\code{z} by \code{1-z}. Apply \code{\link{evalue_correct}} to prevent e-values
of exactly zero.
}
\examples{
# Test hypothesis "greater than uniform distribution in stochastic dominance"

## Hypothesis is violated
z <- rbeta(300, 0.75, 1)
e <- grenander_e(z = z)
max(cumprod(e$e))

## Hypothesis is correct
z <- rbeta(300, 1, 0.75)
e <- grenander_e(z = z)
max(cumprod(e$e))
}
