% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_pit.R
\name{kernel_e}
\alias{kernel_e}
\title{Test for uniformity with kernel density estimator}
\usage{
kernel_e(z, n0 = 10, scalest = "minim", level = 2L, gridsize = 401L, ...)
}
\arguments{
\item{z}{vector of numbers in [0,1].}

\item{n0}{minimum number of observations for starting. All e-values until
\code{n0} (included) are equal to 1, and the density is estimated
starting from the \code{n0+1}th observation.}

\item{scalest}{scale estimator in plug-in bandwidth selection (see
\code{\link[KernSmooth]{dpik}}).}

\item{level}{number of levels of functional estimation in plug-in rule
(see \code{\link[KernSmooth]{dpik}}).}

\item{gridsize}{see \code{\link[KernSmooth]{dpik}}.}

\item{...}{currently not used, only for calling the function with
\code{do.call}.#'}
}
\value{
A list containing the vector of e-values (\code{e}) and a vector of the
bandwidths at the corresponding indices (\code{bws}).
}
\description{
Tests the null hypothesis that \code{z} is uniformly distributed against
any other non-uniform distribution, estimated with a kernel density
estimator.
}
\details{
The density is estimated with
\code{\link[bde]{jonesCorrectionMuller94BoundaryKernel}} and bandwidth
selection uses \code{\link[KernSmooth]{dpik}}. In some cases, this
estimator is not a density (it does not integrate to one), and it is
therefore approximated on the grid 0,0.01,...,0.99,1 and appropriately
rescaled.
}
