% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_misc.R
\name{merge_e}
\alias{merge_e}
\title{Merge e-values and perform hypothesis test}
\usage{
merge_e(e_output)
}
\arguments{
\item{e_output}{output of functions \code{\link{e_pit}},
\code{\link{e_quantile_pit}}, or \code{\link{e_rank_histogram}}.}
}
\value{
A list containing the maximum of the cumulated e-values (\code{e}), the index
where this maximum is attained (\code{max_ind}), and the e-value at the end
of the observation period (\code{e_end}).
}
\description{
Merges a vector of e-values to a single e-value and performs a hypothesis
test (if desired).
}
\details{
If the lag in the e-values is 1, e-values are merged by (cumulative) product.
The null hypothesis can be rejected at level \code{\alpha}if the cumulative
product exceeds the level \code{1/alpha} at least once.

For lag h, all e-values with time lag h are merged by product, and the
average is taken. The null hypothesis can be rejected if this process exceeds
\code{1/alpha} at least once.
}
