% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein_densities.R
\name{mOpt.CN}
\alias{mOpt.CN}
\title{Determine optimal number of weights by CN criterion}
\usage{
mOpt.CN(tData, L)
}
\arguments{
\item{tData}{input data (transformed to fit into the given interval).}

\item{L}{diagonal weight matrix.}
}
\value{
The optimal number of weights (single integer).
}
\description{
Determine optimal number of weights by CN criterion
}
\author{
Code by Bradley C. Turnbull, NC State University, Dr. Sujit K. Ghosh, NC
State University, August 25, 2013.
}
\keyword{internal}
