% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_misc.R
\name{quantile_pit}
\alias{quantile_pit}
\title{Quantile PIT}
\usage{
quantile_pit(y, quantiles, alphas, seed = NULL)
}
\arguments{
\item{y}{observations (numeric vector).}

\item{quantiles}{quantile forecasts. A numeric matrix, the number of rows
should equal the length of \code{y}.}

\item{alphas}{levels of the quantiles. Either a numeric matrix of the same
dimension as \code{quantiles}, giving for each quantile the level
(for example 0.2 for the 20% quantile). If the quantiles are equispaced
(for example, always the 0.1, 0.2, ..., 0.9 or 0.05, 0.10, ..., 0.90,
0.95 quantiles for all forecasts), then \code{alphas} can be specified
as a single number giving the grid spacing of the quantiles (0.1 and
0.05, respectively, in the examples before).}

\item{seed}{seed for random number generation (the PIT are randomized). Can
be omitted.}
}
\value{
A list containing the upper and lower quantile PIT (\code{zu} and \code{zl}).
}
\description{
Computes the quantile PIT for forecasts specified by a collection of
quantiles.
}
