% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sequential_grenander}
\alias{sequential_grenander}
\title{Sequential computation and evaluation of the Grenander estimator}
\usage{
sequential_grenander(z, pos_Z)
}
\arguments{
\item{z}{sorted distinct values of the data \code{Z}.}

\item{pos_Z}{integer vector giving for each index \code{i} the position of
of the i-th data point \code{Z[i]} in \code{z}. Indexing starts at 1.}
}
\value{
Vector of the density evaluated (same length as \code{pos_Z}, the first
entry is set to 1).
}
\description{
For observations \code{Z} and all indices \code{i=2,...,length(Z)}, the
Grenander estimator is computed with data \code{Z[1],...Z[i-1]} and
evaluated at \code{Z[i]}.
}
\keyword{internal}
