% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sequential_ranks}
\alias{sequential_ranks}
\title{Sequential computation of e-values for discrete uniform distribution
(empirical)}
\usage{
sequential_ranks(r, m, n0)
}
\arguments{
\item{r}{observations.}

\item{m}{size of discrete uniform distribution.}

\item{n0}{minimum number of observations for starting. All e-values until
\code{n0} (included) are equal to 1.}

\item{...}{currently not used, only for calling the function with
\code{do.call}.}
}
\value{
List containing the of e-values.
}
\description{
For a sample \code{r} in \code{1,...,m}, compute e-values for testing the
null hypothesis that \code{r} follows the uniform distribution on
\code{1,...,m}.
}
\keyword{internal}
