% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_example.R
\name{simulate_pit}
\alias{simulate_pit}
\title{Simulation example for testing the functions}
\usage{
simulate_pit(n, type, bias, dispersion, K = NULL, seed = NULL)
}
\arguments{
\item{n}{number of simulations.}

\item{type}{\code{"pit"} for the usual PIT, \code{"quantile_pit"} for the
quantile PIT, \code{"rank_histogram"} for the ensemble rank histograms.}

\item{bias}{the bias parameter (single number).}

\item{dispersion}{the dispersion error (single number greater than -1).}

\item{K}{number of equispaced quantiles for \code{type="quantile_pit"}, or
ensemble size for \code{type="rank_histogram"}. Has no effect for the
usual PIT.}

\item{seed}{seed for random number generation (can be omitted).}
}
\value{
A list containing the pit \code{z} for the \code{type="pit"}, the upper and
lower quantile pit \code{zu,zl} for \code{type="quantile_pit"} or the
ranks \code{r} for \code{type="rank_histogram"}.
}
\description{
Generates observations \code{y} following a standard normal distribution,
and the probabilistic forecast is normal with mean \code{0+bias} and
variance \code{1+(dispersion error)}.
}
\details{
For the quantile PIT, the \code{K} equispaced quantiles are extracted from
the forecast distribution (for example, the 0.1,0.2,...0.9 quantiles for
\code{K=9}). For the rank histograms, \code{K} ensemble forecasts are
generated randomly from the forecast distribution.
}
