% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein_densities.R
\name{solveWeights}
\alias{solveWeights}
\title{Compute weight vector}
\usage{
solveWeights(m, Fn, lower, upper, Dmat, dvec, monotone, settings)
}
\arguments{
\item{m}{maximum degree in Bernstein polynomial.}

\item{Fn}{ecdf of transformed data.}

\item{lower, upper}{lower and upper bound of support.}

\item{Dmat, dvec}{matrix and vector for QP problem.}

\item{monotone}{direction of monotonicity.}

\item{settings}{settings for \code{\link[osqp]{solve_osqp}}.}
}
\value{
Vector of weights.
}
\description{
Compute weight vector
}
\author{
Original code by Bradley C. Turnbull, NC State University, Dr. Sujit K.
Ghosh, NC State University, August 25, 2013, suitably adapted by Sebastian
Arnold and Alexander Henzi.
}
\keyword{internal}
