% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{strigammav}
\alias{strigammav}
\title{Wrapper for sums of trigamma function}
\arguments{
\item{u}{numeric vector (NumericVector).}

\item{v}{numeric vector (NumericVector, same length as \code{u})}
}
\value{
Single number (double), the result.
}
\description{
Apply trigamma function to vector and compute scalar product with another
vector
}
\keyword{internal}
