% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_rank_histogram.R
\name{empirical_e}
\alias{empirical_e}
\title{Test discrete uniform distribution against empirical distribution}
\usage{
empirical_e(r, m, n0 = 20, ...)
}
\arguments{
\item{r}{observations.}

\item{m}{size of discrete uniform distribution.}

\item{n0}{minimum number of observations for starting. All e-values until
\code{n0} (included) are equal to 1.}

\item{...}{currently not used, only for calling the function with
\code{do.call}.}
}
\value{
List containing the e-values.
}
\description{
Tests the null hypothesis that \code{r} are generated from a discrete
uniform distribution, against the alternative defined by their empirical
distribution.
}
\examples{
r <- simulate_pit(1000, "rank_histogram", K = 20, bias = 0.2, dispersion = 0)
e <- empirical_e(r = r$r, m = 20)
max(cumprod(e$e))
}
