% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_misc.R
\name{quantile_pit}
\alias{quantile_pit}
\title{Quantile PIT}
\usage{
quantile_pit(y, quantiles, seed = NULL, check = FALSE)
}
\arguments{
\item{y}{observations (numeric vector).}

\item{quantiles}{quantile forecasts. A numeric matrix, the number of rows
should equal the length of \code{y}. It is assumed that the \code{j}th
column contains the \code{j/(K+1)}th quantile, where
\code{K=ncol(quantiles)}.}

\item{seed}{seed for random number generation (the PITs are randomized). Can
be omitted.}

\item{check}{check for correct format of input parameters.}
}
\value{
A list containing the upper and lower quantile PIT (\code{zu} and \code{zl}).
}
\description{
Computes the quantile PIT for forecasts specified by a collection of
quantiles.
}
